Fix workspace directory resolution for consistent container naming
This change ensures all devgo commands use the same workspace directory
and container naming logic, regardless of the subdirectory they're run from.

Previously, when running `devgo up` from a workspace root and `devgo shell`
from a subdirectory, they would use different container names:
- `devgo up` from ~/ros1 → container: devgo-ros1
- `devgo shell` from ~/ros1/src/turtlesim_cleaner → container: devgo-turtlesim_cleaner

Now all commands consistently use the directory containing devcontainer.json
as the workspace directory for container naming.

Changes:
- Move determineWorkspaceFolder() and determineContainerName() to root.go
- Update workspace folder logic to use devcontainer.json location
- Apply consistent workspace resolution across all commands (up, shell, exec, down, stop)
- Update tests to work with new function signature
