+++ cmd/stop.go
@@ -12,12 +12,12 @@ import (
 )
 
 func runStopCommand(args []string) error {
-	workspaceDir := determineWorkspaceFolder()
-	
 	devcontainerPath, err := findDevcontainerConfig("")
 	if err != nil {
 		return fmt.Errorf("failed to find devcontainer config: %w", err)
 	}
+	
+	workspaceDir := determineWorkspaceFolder(devcontainerPath)
 
 	devContainer, err := devcontainer.Parse(devcontainerPath)
 	if err != nil {
